BROKEN BUILD SERIES: Upstream 0e2029a Jul 11

+++ qemu/target-i386/translate.c
@@ -2016,7 +2016,7 @@ static void gen_lea_modrm(DisasContext *s, int modrm, int *reg_ptr, int *offset_
             break;
         default:
         case 2:
-            disp = ldl_code(s->pc);
+            disp = (int32_t)ldl_code(s->pc);
             s->pc += 4;
             break;
         }
@@ -5714,7 +5714,7 @@ static target_ulong disas_insn(DisasContext *s, target_ulong pc_start)
 			case 0x3b:
 
 #endif
-				switch(op & 7) {
+                switch(op & 7) {
                 case 0:
                     switch(op >> 4) {
                     case 0:
